backend/requirements.txt
# V6rge Backend - Complete Dependencies
# Install PyTorch first: pip install torch==2.3.1 torchvision==0.18.1 torchaudio==2.3.1 --index-url https://download.pytorch.org/whl/cu121
# Then: pip install -r requirements.txt

# Transformers ecosystem
transformers>=4.44.0
accelerate>=0.33.0
diffusers>=0.30.0
safetensors>=0.4.2
sentencepiece>=0.2.0
tokenizers>=0.19.1
huggingface-hub>=0.25.0
peft>=0.17.0
# autoawq removed (incompatible with torch>=2.6)

# Vision models
qwen-vl-utils>=0.0.5

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy<2.0.0,>=1.26.0
scipy>=1.13.0
scikit-image>=0.22.0

# Image tools - RealESRGAN stack
realesrgan>=0.3.0
basicsr>=1.4.2
gfpgan>=1.3.8
facexlib>=0.3.0

# Background removal
rembg>=2.0.50

# Audio processing
soundfile>=0.12.1
librosa>=0.10.1

# Audio separation
audio-separator>=0.17.0

# Web server
flask>=3.0.0
flask-cors>=4.0.0

# 3D generation
trimesh>=4.0.0
pygltflib>=1.16.0

# Utils
omegaconf>=2.3.0
einops>=0.7.0
pyyaml>=6.0.1
tqdm>=4.66.0

# TTS - Chatterbox Turbo
chatterbox-tts
conformer
s3tokenizer
resemble-perth
pyloudnorm
